# backy/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Home API",
    "description": "Comprehensive REST API for the Smart Home property management platform. Covers Authentication, User & Profile management, Property listings, Payments, Wishlists, Activity Logs, Addresses, Carousels, and Currency.",
    "version": "1.0.0",
    "contact": {
      "email": "support@smarthome.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.smarthome.com",
      "description": "Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide your JWT access token. Obtain one via POST /api/auth/login."
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT token"
      },
      "Forbidden": {
        "description": "Insufficient role permissions"
      },
      "BadRequest": {
        "description": "Validation error or missing required fields"
      },
      "NotFound": {
        "description": "Resource not found"
      }
    }
  },
  "paths": {
...
    "/api/profile/{id}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get profile by user ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile returned"
          }
        }
      },
      "delete": {
        "tags": [
          "Profile"
        ],
        "summary": "Delete a profile (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile deleted"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/properties/{id}": {
      "get": {
        "tags": [
          "Property"
        ],
        "summary": "Get property by ID (public)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property data"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Property"
        ],
        "summary": "Update a property (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "pricing": {
                    "type": "string",
                    "description": "JSON string"
                  },
                  "features": {
                    "type": "string",
                    "description": "JSON string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property updated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Property"
        ],
        "summary": "Delete a property (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property deleted"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/properties/": {
      "post": {
        "tags": [
          "Property"
        ],
        "summary": "Create a new property listing (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "listingType",
                  "propertyType"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "listingType": {
                    "type": "string",
                    "enum": [
                      "for_rent",
                      "for_sale"
                    ]
                  },
                  "propertyType": {
                    "type": "string",
                    "enum": [
                      "residential",
                      "commercial",
                      "land"
                    ]
                  },
                  "propertySubType": {
                    "type": "string"
                  },
                  "pricing": {
                    "type": "string",
                    "description": "JSON: { rentPrice, salePrice, currency }"
                  },
                  "features": {
                    "type": "string",
                    "description": "JSON array of feature strings"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property created"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/payments/paystack/webhook": {
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "Paystack webhook endpoint (internal)",
        "description": "Receives and processes Paystack event webhooks. Secured via Paystack-Signature header validation.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed"
          }
        }
      }
    },
...
    "/api/wishlist/": {
      "get": {
        "tags": [
          "Wishlist"
        ],
        "summary": "Get current user's wishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Wishlist with property details"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/addresses/": {
      "post": {
        "tags": [
          "Address"
        ],
        "summary": "Create a new address",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "street",
                  "city",
                  "state",
                  "country"
                ],
                "properties": {
                  "street": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "postalCode": {
                    "type": "string"
                  },
                  "isDefault": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Address created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "get": {
        "tags": [
          "Address"
        ],
        "summary": "Get all addresses for current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Addresses list"
          }
        }
      }
    },
...
    "/api/addresses/{id}": {
      "patch": {
        "tags": [
          "Address"
        ],
        "summary": "Update an address",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "street": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "isDefault": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address updated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "delete": {
        "tags": [
          "Address"
        ],
        "summary": "Delete an address",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/carousels/": {
      "post": {
        "tags": [
          "Carousel"
        ],
        "summary": "Create a carousel banner (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "subtitle": {
                    "type": "string"
                  },
                  "link": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Banner created"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "tags": [
          "Carousel"
        ],
        "summary": "List all carousel banners (public)",
        "responses": {
          "200": {
            "description": "All banners"
          }
        }
      }
    },
...
    "/api/carousels/{bannerId}": {
      "patch": {
        "tags": [
          "Carousel"
        ],
        "summary": "Update a carousel banner (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bannerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "subtitle": {
                    "type": "string"
                  },
                  "link": {
                    "type": "string"
                  },
                  "active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Banner updated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/carousels/{id}": {
      "delete": {
        "tags": [
          "Carousel"
        ],
        "summary": "Delete a carousel banner (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Banner deleted"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/currency-code/": {
      "post": {
        "tags": [
          "Currency"
        ],
        "summary": "Add a supported currency (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "symbol",
                  "name"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "NGN"
                  },
                  "symbol": {
                    "type": "string",
                    "example": "₦"
                  },
                  "name": {
                    "type": "string",
                    "example": "Nigerian Naira"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Currency added"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "tags": [
          "Currency"
        ],
        "summary": "List all supported currencies (public)",
        "responses": {
          "200": {
            "description": "Currency list"
          }
        }
      }
    },
    "/api/currency-code/{id}": {
      "delete": {
        "tags": [
          "Currency"
        ],
        "summary": "Remove a currency (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Currency removed"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  }
}